requirements.txt
streamlit==1.32.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.14.0
keras-cv==0.9.0
pillow==10.0.1
gspread==5.11.3
oauth2client==4.1.3
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.0
nltk==3.8.1
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
PyAudio==0.2.13
Flask==2.3.3
